requirements.txt
functions-framework==3.*
Flask==2.*
google-cloud-storage==2.*
